# api/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.CourseResponse:
    properties:
      capacity:
        type: integer
      code:
        type: string
      course_times:
        items:
          $ref: '#/definitions/dto.CourseTimeResponse'
        type: array
      created_at:
        type: string
      exam_end:
        type: string
      exam_start:
        type: string
      faculty_id:
        type: string
      faculty_name_en:
        type: string
      faculty_name_fa:
        type: string
      gender_restriction:
        type: string
      id:
        type: string
      name:
        type: string
      professor_id:
        type: string
      professor_name:
        type: string
      semester_id:
        type: string
      university_id:
        type: string
      updated_at:
        type: string
      weight:
        type: integer
    type: object
  dto.CourseTimeResponse:
    properties:
      course_id:
        type: string
      day_of_week:
        type: integer
      end_time:
        type: string
      id:
        type: string
      start_time:
        type: string
    type: object
...
  dto.ErrorResponse:
    properties:
      error:
        description: |-
          error message
          example: invalid request format
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  title: Termustat API
  version: "1.0"
paths:
...
  /v1/user/courses/select:
    post:
      consumes:
      - application/json
      description: Adds a course to the current user's schedule
      parameters:
      - description: course_id and semester_id
        in: body
        name: body
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 'message: Course added successfully'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Course or semester not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "409":
          description: Conflict (e.g. already selected)
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Select Course
      tags:
      - user-courses
  /v1/user/courses/select/{courseId}:
    delete:
      description: Removes a course from the current user's schedule
      parameters:
      - description: Course ID
        in: path
        name: courseId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 'message: Course removed successfully'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid course ID
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Course not found in user's schedule
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Remove Course
      tags:
      - user-courses
  /v1/user/courses/selected:
    get:
      description: Retrieves all courses selected by the user for a given semester
      parameters:
      - description: Semester ID
        in: query
        name: semester_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.CourseResponse'
            type: array
        "400":
          description: Invalid semester ID
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get User Courses
      tags:
      - user-courses
  /v1/user/courses/validate:
    get:
      description: Checks if adding the selected course causes any time conflicts
      parameters:
      - description: Course ID
        in: query
        name: course_id
        required: true
        type: string
      - description: Semester ID
        in: query
        name: semester_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 'message: No time conflicts found'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "409":
          description: Time conflict exists
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Validate Time Conflicts
      tags:
      - user-courses
...
swagger: "2.0"